# Makes computing visible rows faster