--- a/src/components/shared/TreeView.js
+++ b/src/components/shared/TreeView.js
@@ -316,6 +316,7 @@ class TreeView<
 > extends React.PureComponent<TreeViewProps<NodeIndex, DisplayData>> {
   _specialItems: (NodeIndex | null)[];
   _visibleRows: NodeIndex[];
+  _expandedNodes: Set<NodeIndex | null>;
   _list: VirtualList | null;
   _takeListRef = (list: VirtualList | null) => (this._list = list);
 
@@ -327,6 +328,7 @@ class TreeView<
     (this: any)._onCopy = this._onCopy.bind(this);
     (this: any)._onRowClicked = this._onRowClicked.bind(this);
     this._specialItems = [props.selectedNodeId];
+    this._expandedNodes = new Set(props.expandedNodeIds);
     this._visibleRows = this._getAllVisibleRows(props);
     this._list = null;
   }
@@ -349,6 +351,7 @@ class TreeView<
       nextProps.tree !== this.props.tree ||
       nextProps.expandedNodeIds !== this.props.expandedNodeIds
     ) {
+      this._expandedNodes = new Set(nextProps.expandedNodeIds);
       this._visibleRows = this._getAllVisibleRows(nextProps);
     }
   }
@@ -356,7 +359,6 @@ class TreeView<
   _renderRow(nodeId: NodeIndex, index: number, columnIndex: number) {
     const {
       tree,
-      expandedNodeIds,
       fixedColumns,
       mainColumn,
       appendageColumn,
@@ -380,7 +382,7 @@ class TreeView<
       );
     }
     const canBeExpanded = tree.hasChildren(nodeId);
-    const isExpanded = expandedNodeIds.includes(nodeId);
+    const isExpanded = !this._isCollapsed(nodeId);
     return (
       <TreeViewRowScrolledColumns
         displayData={displayData}
@@ -408,7 +410,7 @@ class TreeView<
     depth: number
   ) {
     arr.push(nodeId);
-    if (!props.expandedNodeIds.includes(nodeId)) {
+    if (this._isCollapsed(nodeId)) {
       return;
     }
     const children = props.tree.getChildren(nodeId);
@@ -429,7 +431,7 @@ class TreeView<
   }
 
   _isCollapsed(nodeId: NodeIndex): boolean {
-    return !this.props.expandedNodeIds.includes(nodeId);
+    return !this._expandedNodes.has(nodeId);
   }
 
   _toggle(
@@ -437,7 +439,7 @@ class TreeView<
     newExpanded: boolean = this._isCollapsed(nodeId),
     toggleAll: * = false
   ) {
-    const newSet = new Set(this.props.expandedNodeIds);
+    const newSet = new Set(this._expandedNodes);
     if (newExpanded) {
       newSet.add(nodeId);
       if (toggleAll) {