The testcase is this profile: http://localhost:4242/public/280e10708c0660695ef5af145c53179d29d9796c/calltree/?hiddenThreads=&react_perf&thread=0&threadOrder=0-2-3-4-5-1&v=3

STR is:
1. Right click on "root"
2. Click "Expand All"
=> On my computer, this STR goes from 11s to 5s with this patch.
3. Then collapse a node
=> On my computer, this STR goes from 4.4s to 3.9s.