@julienw 2018-03-13
> Maybe I didn't want the overhead of creating a set in componentWillReceiveProps every time something changes.

Yeah, the overhead is small for small data, and likely bigger for bigger data, yet the algorithmic advantage we get from a set is also bigger in this case and exceeds the penalty.